requirements.txt
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Biology & Chemistry
biopython>=1.81

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Data Processing
h5py>=3.8.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0